iommu/omap: change object name to use device name
The iommu objecy name is currently assigned to be the OF
name to support DT iommu devices. This is probably done to
allow the OMAP3 ISP node to preserve its name, but the DT
nodes are typically named 'mmu@<address>' and so will have
a common OF name for multiple nodes. The OMAP3 ISP instead
should be changed to use the DT device name, so that there
is no ambiguity. This also facilitates an easier cleanup
of the non-DT dependencies in the driver.

Signed-off-by: Suman Anna <s-anna@ti.com>