ARM: dts: OMAP: Add hwlock-base-id property to hwspinlock nodes
Add the generic property "hwlock-base-id" to the hwspinlock DT nodes
on all applicable OMAP SoCS (OMAP4, OMAP5, DRA7, AM33xx and AM43xx).
This common property is required by the hwspinlock core to be able
to translate client locks properly. All the current OMAP SoCs only
have a single instance of the IP, and so will use a common value of 0.

Signed-off-by: Suman Anna <s-anna@ti.com>