hwspinlock/core: add device tree support
This patch adds a new OF-friendly API for hwspinlock clients
to use/request locks from a hwspinlock device instantiated
through a device-tree blob. The of_hwspin_lock_get_id() API
can be used by hwspinlock clients to get the id for a specific
lock using the phandle + args specifier, so that it can be
requested using the available hwspin_lock_request_specific()
API.

NOTE:
The logic used is based on an implicit usage of a single
hwspinlock bank/device. This will need to be revised when
multiple banks show up.

Signed-off-by: Suman Anna <s-anna@ti.com>