remoteproc/pru: fix possible erroneous intc unconfig during shutdown
The intc_config structure elements are currently not initialized
specifically in probe, and they are filled in either during the
pru_rproc_get() if a client DT node has the ti,pru-interrupt-map
property set or from the firmware's resource table vendor custom
interrupt resource type while loading/starting a firmware. It is
possible that a client neither has the DT property nor uses the
custom resource in its firmware, and they are left initialized to
their initial allocated values. This may result in having some
valid mappings and resulting in a possible erroneous INTC unconfig
while shutting down a processor.

Fix this by introducing a boolean variable and using it to invoke
invoke the pruss_intc_unconfigure() conditionally. This flag is
initialized only if the firmware has a PRU custom vendor interrupt
resource in its resource table.

Reported-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>