ARM: dts: da850: Add the PRU-ICSS DT nodes
Add the DT nodes for the PRU-ICSS on OMAP-L13x/AM18x/DA8xx family
of SoCs. These SoCs contain a single PRU-ICSS instance and is
represented by the pruss-soc-bus node and a child PRUSS node.
PRU-ICSS is not supported on AM1802 SoC though in the AM33xx
family, so the nodes are added in disabled state to the common
da850 dtsi file. They should be enabled in only those derivative
board files that use a SoC containing PRU-ICSS.

The PRUSS subsystem node contains the entire address space and
the various interrupts generated towards the main MPU. The various
sub-modules of the PRU-ICSS are represented as individual child
nodes (so platform devices themselves) of the PRUSS subsystem node.
These include the two PRU cores and the interrupt controller. All
the Data RAMs are represented within a child node of its own named
'memories' without any compatible.

The DT nodes use all standard properties. The regs property in
the PRU nodes define the addresses for the Instruction RAM, the
Debug and Control sub-modules for that PRU core. The firmware for
each PRU core is defined through a 'firmware-name' property.

The default names for the firmware images for each PRU core are
defined as follows (these can be adjusted either in derivative
board dts files or through sysfs at runtime if required):
     PRU-ICSS PRU0 Core: da850-pru0-fw
     PRU-ICSS PRU1 Core: da850-pru1-fw

Signed-off-by: Suman Anna <s-anna@ti.com>