ARM: dts: da850:: Add PRU system events for virtio
Two PRU system events "vring" and "kick" have been added to each
of the PRU nodes in the PRU-ICSS remote processor subsystem to
enable the virtio/rpmsg communication between MPU and that PRU
core. The additions are done in the base da850.dtsi file, and
so are inherited by all the DA850/AM1808/OMAP-L138 derivative
boards. Do note that PRUSS is not available on all AM18xx SoCs.
The PRU system event numbers are chosen starting from the first
non-external peripheral sourced event numbers and can be adjusted
in the respective board file as per need.

Signed-off-by: Suman Anna <s-anna@ti.com>