dt-bindings: remoteproc: ti-pruss: Update bindings to add MII_G_RT info
The ICSSG IP on K3 SoCs has couple of additional sub-modules w.r.t previous
generation PRU-ICSS IPs. Update the bindings to add the details for one
such sub-module - MII_G_RT.

While at this, update all the existing AM6x references to AM65x to reflect
the proper SoC name to be used.

NOTE: The usage of the second IEP and MII_RT sub-module instances are yet
to be determined, and so are yet to be added.

Fixes: b96ed2ff575f ("dt-bindings: remoteproc: ti-pruss: Update bindings for K3 AM6x SoCs")
Signed-off-by: Suman Anna <s-anna@ti.com>